refactor(api): rename http::core to http::router
