+++ packages/api/core/src/http/mod.rs
@@ -1,6 +1,6 @@
 mod context;
 pub mod controllers;
-mod core;
+mod router;
 mod error;
 mod errors;
 mod extractors;
@@ -8,10 +8,10 @@ pub mod middleware;
 pub mod resources;
 pub mod response;
 
-pub use self::core::init;
+pub use self::router::init;
 pub use context::Context;
 pub use error::Error as Error;
 pub use response::JsonResponse;
 
 #[cfg(test)]
-pub use self::core::router;
+pub use self::router::router;
